pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "autoredocs"
version = "0.3.0"
description = "Real-time, self-maintaining code documentation CLI tool"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "Autoredocs Team" },
]
keywords = ["documentation", "cli", "code-docs", "auto-documentation", "docstring"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Documentation",
    "Topic :: Software Development :: Documentation",
]
dependencies = [
    "typer[all]>=0.9.0",
    "rich>=13.0.0",
    "watchfiles>=0.21.0",
    "jinja2>=3.1.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
ai = [
    "openai>=1.0",
    "python-dotenv>=1.0",
]
server = [
    "fastapi>=0.100.0",
    "uvicorn[standard]>=0.20.0",
]
deploy = [
    "httpx>=0.25.0",
    "boto3>=1.28.0",
]
all = [
    "autoredocs[ai,server,deploy]",
]
dev = [
    "pytest>=8.0",
    "pytest-cov>=4.0",
    "openai>=1.0",
    "python-dotenv>=1.0",
    "groq>=0.4.0",
]

[project.scripts]
autoredocs = "autoredocs.cli:app"

[project.urls]
Homepage = "https://github.com/ShivamBwaj/autoredocs"
Documentation = "https://shivambwaj.github.io/autoredocs/"
Repository = "https://github.com/ShivamBwaj/autoredocs"
Issues = "https://github.com/ShivamBwaj/autoredocs/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/autoredocs"]

[tool.pytest.ini_options]
minversion = "8.0"
addopts = "-ra --tb=short"
testpaths = ["tests"]

[tool.ruff]
target-version = "py310"
line-length = 100
exclude = ["tests/fixtures/bad_syntax.py"]
